Created a reutilizable error-form-Validators
i made this to make the HTML register code cleaner and making it easier to debug. Also, added some others validations, and implemented it in the register HTML and in the register.ts, importing it directly, and not injecting in the constructor, making the form error STANDALONE, great way to clean the constructor.